requirements.txt
gspread 
google-auth